+++ client.go
@@ -1446,6 +1446,13 @@ func validateQueueName(queueName string) error {
 	return nil
 }
 
+// JobListResult is the result of a job list operation. It contains a list of
+// jobs and a cursor for fetching the next page of results.
+type JobListResult struct {
+	Jobs       []*rivertype.JobRow
+	LastCursor *JobListCursor
+}
+
 // JobList returns a paginated list of jobs matching the provided filters. The
 // provided context is used for the underlying Postgres query and can be used to
 // cancel the operation or apply a timeout.
@@ -1455,7 +1462,7 @@ func validateQueueName(queueName string) error {
 //	if err != nil {
 //		// handle error
 //	}
-func (c *Client[TTx]) JobList(ctx context.Context, params *JobListParams) ([]*rivertype.JobRow, error) {
+func (c *Client[TTx]) JobList(ctx context.Context, params *JobListParams) (*JobListResult, error) {
 	if !c.driver.HasPool() {
 		return nil, errNoDriverDBPool
 	}
@@ -1468,7 +1475,15 @@ func (c *Client[TTx]) JobList(ctx context.Context, params *JobListParams) ([]*ri
 		return nil, err
 	}
 
-	return dblist.JobList(ctx, c.driver.GetExecutor(), dbParams)
+	jobs, err := dblist.JobList(ctx, c.driver.GetExecutor(), dbParams)
+	if err != nil {
+		return nil, err
+	}
+	res := &JobListResult{Jobs: jobs}
+	if len(jobs) > 0 {
+		res.LastCursor = JobListCursorFromJob(jobs[len(jobs)-1], params.sortField)
+	}
+	return res, nil
 }
 
 // JobListTx returns a paginated list of jobs matching the provided filters. The
@@ -1480,16 +1495,25 @@ func (c *Client[TTx]) JobList(ctx context.Context, params *JobListParams) ([]*ri
 //	if err != nil {
 //		// handle error
 //	}
-func (c *Client[TTx]) JobListTx(ctx context.Context, tx TTx, params *JobListParams) ([]*rivertype.JobRow, error) {
+func (c *Client[TTx]) JobListTx(ctx context.Context, tx TTx, params *JobListParams) (*JobListResult, error) {
 	if params == nil {
 		params = NewJobListParams()
 	}
+
 	dbParams, err := params.toDBParams()
 	if err != nil {
 		return nil, err
 	}
 
-	return dblist.JobList(ctx, c.driver.UnwrapExecutor(tx), dbParams)
+	jobs, err := dblist.JobList(ctx, c.driver.UnwrapExecutor(tx), dbParams)
+	if err != nil {
+		return nil, err
+	}
+	res := &JobListResult{Jobs: jobs}
+	if len(jobs) > 0 {
+		res.LastCursor = JobListCursorFromJob(jobs[len(jobs)-1], params.sortField)
+	}
+	return res, nil
 }
 
 // PeriodicJobs returns the currently configured set of periodic jobs for the