+++ job_list_params.go
@@ -15,19 +15,34 @@ import (
 // JobListCursor is used to specify a starting point for a paginated
 // job list query.
 type JobListCursor struct {
-	id    int64
-	kind  string
-	queue string
-	time  time.Time
+	id        int64
+	kind      string
+	queue     string
+	sortField JobListOrderByField
+	time      time.Time
 }
 
 // JobListCursorFromJob creates a JobListCursor from a JobRow.
-func JobListCursorFromJob(job *rivertype.JobRow) *JobListCursor {
+func JobListCursorFromJob(job *rivertype.JobRow, sortField JobListOrderByField) *JobListCursor {
+	time := job.CreatedAt
+	switch sortField {
+	case JobListOrderByTime:
+		time = jobListTimeValue(job)
+	case JobListOrderByFinalizedAt:
+		if job.FinalizedAt != nil {
+			time = *job.FinalizedAt
+		}
+	case JobListOrderByScheduledAt:
+		time = job.ScheduledAt
+	default:
+		panic("invalid sort field")
+	}
 	return &JobListCursor{
-		id:    job.ID,
-		kind:  job.Kind,
-		queue: job.Queue,
-		time:  jobListTimeValue(job),
+		id:        job.ID,
+		kind:      job.Kind,
+		queue:     job.Queue,
+		sortField: sortField,
+		time:      time,
 	}
 }
 
@@ -46,10 +61,11 @@ func (c *JobListCursor) UnmarshalText(text []byte) error {
 		return err
 	}
 	*c = JobListCursor{
-		id:    wrapperValue.ID,
-		kind:  wrapperValue.Kind,
-		queue: wrapperValue.Queue,
-		time:  wrapperValue.Time,
+		id:        wrapperValue.ID,
+		kind:      wrapperValue.Kind,
+		queue:     wrapperValue.Queue,
+		sortField: JobListOrderByField(wrapperValue.SortField),
+		time:      wrapperValue.Time,
 	}
 	return nil
 }
@@ -58,10 +74,11 @@ func (c *JobListCursor) UnmarshalText(text []byte) error {
 // opaque string.
 func (c JobListCursor) MarshalText() ([]byte, error) {
 	wrapperValue := jobListPaginationCursorJSON{
-		ID:    c.id,
-		Kind:  c.kind,
-		Queue: c.queue,
-		Time:  c.time,
+		ID:        c.id,
+		Kind:      c.kind,
+		Queue:     c.queue,
+		SortField: string(c.sortField),
+		Time:      c.time,
 	}
 	data, err := json.Marshal(wrapperValue)
 	if err != nil {
@@ -73,10 +90,11 @@ func (c JobListCursor) MarshalText() ([]byte, error) {
 }
 
 type jobListPaginationCursorJSON struct {
-	ID    int64     `json:"id"`
-	Kind  string    `json:"kind"`
-	Queue string    `json:"queue"`
-	Time  time.Time `json:"time"`
+	ID        int64     `json:"id"`
+	Kind      string    `json:"kind"`
+	Queue     string    `json:"queue"`
+	SortField string    `json:"sort_field"`
+	Time      time.Time `json:"time"`
 }
 
 // SortOrder specifies the direction of a sort.
@@ -90,12 +108,20 @@ const (
 )
 
 // JobListOrderByField specifies the field to sort by.
-type JobListOrderByField int
+type JobListOrderByField string
 
 const (
+	// JobListOrderByFinalizedAt specifies that the sort should be by
+	// finalized_at.
+	//
+	// This option must be used in conjunction with filtering by only finalized
+	// job states.
+	JobListOrderByFinalizedAt JobListOrderByField = "finalized_at"
+	// JobListOrderByScheduledAt specifies that the sort should be by scheduled_at.
+	JobListOrderByScheduledAt JobListOrderByField = "scheduled_at"
 	// JobListOrderByTime specifies that the sort should be by time. The specific
-	// time field used will vary by job state.
-	JobListOrderByTime JobListOrderByField = iota
+	// time field used will vary by the first specified job state.
+	JobListOrderByTime JobListOrderByField = "time"
 )
 
 // JobListParams specifies the parameters for a JobList query. It must be
@@ -107,11 +133,12 @@ type JobListParams struct {
 	after            *JobListCursor
 	kinds            []string
 	metadataFragment string
+	overrodeState    bool
 	paginationCount  int32
 	queues           []string
 	sortField        JobListOrderByField
 	sortOrder        SortOrder
-	state            rivertype.JobState
+	states           []rivertype.JobState
 }
 
 // NewJobListParams creates a new JobListParams to return available jobs sorted
@@ -121,7 +148,15 @@ func NewJobListParams() *JobListParams {
 		paginationCount: 100,
 		sortField:       JobListOrderByTime,
 		sortOrder:       SortOrderAsc,
-		state:           rivertype.JobStateAvailable,
+		states: []rivertype.JobState{
+			JobStateAvailable,
+			JobStateCancelled,
+			JobStateCompleted,
+			JobStateDiscarded,
+			JobStateRetryable,
+			JobStateRunning,
+			JobStateScheduled,
+		},
 	}
 }
 
@@ -130,11 +165,12 @@ func (p *JobListParams) copy() *JobListParams {
 		after:            p.after,
 		kinds:            append([]string(nil), p.kinds...),
 		metadataFragment: p.metadataFragment,
+		overrodeState:    p.overrodeState,
 		paginationCount:  p.paginationCount,
 		queues:           append([]string(nil), p.queues...),
 		sortField:        p.sortField,
 		sortOrder:        p.sortOrder,
-		state:            p.state,
+		states:           append([]rivertype.JobState(nil), p.states...),
 	}
 }
 
@@ -154,10 +190,30 @@ func (p *JobListParams) toDBParams() (*dblist.JobListParams, error) {
 		return nil, errors.New("invalid sort order")
 	}
 
-	if p.sortField != JobListOrderByTime {
-		return nil, errors.New("invalid sort field")
+	if p.sortField == JobListOrderByFinalizedAt {
+		currentNonFinalizedStates := make([]rivertype.JobState, 0, len(p.states))
+		for _, state := range p.states {
+			//nolint:exhaustive
+			switch state {
+			case JobStateCancelled, JobStateCompleted, JobStateDiscarded:
+			default:
+				currentNonFinalizedStates = append(currentNonFinalizedStates, state)
+			}
+		}
+		// This indicates the user overrode the States list with only non-finalized
+		// states prior to then requesting FinalizedAt ordering.
+		if len(currentNonFinalizedStates) == 0 {
+			return nil, fmt.Errorf("cannot order by finalized_at with non-finalized state filters %+v", currentNonFinalizedStates)
+		}
 	}
-	timeField := jobListTimeFieldForState(p.state)
+
+	var timeField string
+	if len(p.states) > 0 && p.sortField == JobListOrderByTime {
+		timeField = jobListTimeFieldForState(p.states[0])
+	} else {
+		timeField = string(p.sortField)
+	}
+
 	orderBy = append(orderBy, []dblist.JobListOrderBy{
 		{Expr: timeField, Order: sortOrder},
 		{Expr: "id", Order: sortOrder},
@@ -193,7 +249,7 @@ func (p *JobListParams) toDBParams() (*dblist.JobListParams, error) {
 		OrderBy:    orderBy,
 		Priorities: nil,
 		Queues:     p.queues,
-		State:      p.state,
+		States:     p.states,
 	}
 
 	return dbParams, nil
@@ -249,18 +305,38 @@ func (p *JobListParams) Queues(queues ...string) *JobListParams {
 
 // OrderBy returns an updated filter set that will sort the results using the
 // specified field and direction.
+//
+// If ordering by FinalizedAt, the States filter will be set to only include
+// finalized job states unless it has already been overridden.
 func (p *JobListParams) OrderBy(field JobListOrderByField, direction SortOrder) *JobListParams {
 	result := p.copy()
+	switch field {
+	case JobListOrderByTime, JobListOrderByScheduledAt:
+		result.sortField = field
+	case JobListOrderByFinalizedAt:
+		result.sortField = field
+		if !p.overrodeState {
+			result.states = []rivertype.JobState{
+				JobStateCancelled,
+				JobStateCompleted,
+				JobStateDiscarded,
+			}
+		}
+	default:
+		panic("invalid order by field")
+	}
 	result.sortField = field
 	result.sortOrder = direction
 	return result
 }
 
-// State returns an updated filter set that will only return jobs in the given
-// state.
-func (p *JobListParams) State(state rivertype.JobState) *JobListParams {
+// States returns an updated filter set that will only return jobs in the given
+// states.
+func (p *JobListParams) States(states ...rivertype.JobState) *JobListParams {
 	result := p.copy()
-	result.state = state
+	result.states = make([]rivertype.JobState, len(states))
+	result.overrodeState = true
+	copy(result.states, states)
 	return result
 }